[SPARK-46268][PS][CONNECT][TESTS] Re-organize `StatsTests`
### What changes were proposed in this pull request?
Re-organize `StatsTests`

### Why are the changes needed?
break the big test file by grouping test cases by topics

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44185 from zhengruifeng/ps_reorg_test_stats.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
