+++ python/pyspark/pandas/tests/computation/test_stats.py
@@ -15,13 +15,10 @@
 # limitations under the License.
 #
 
-import unittest
-
 import numpy as np
 import pandas as pd
 
 from pyspark import pandas as ps
-from pyspark.pandas.config import option_context
 from pyspark.testing.pandasutils import PandasOnSparkTestCase, SPARK_CONF_ARROW_ENABLED
 from pyspark.testing.sqlutils import SQLTestUtils
 
@@ -160,99 +157,6 @@ class StatsTestsMixin:
         ):
             psdf.D.abs()
 
-    def test_axis_on_dataframe(self):
-        # The number of each count is intentionally big
-        # because when data is small, it executes a shortcut.
-        # Less than 'compute.shortcut_limit' will execute a shortcut
-        # by using collected pandas dataframe directly.
-        # now we set the 'compute.shortcut_limit' as 1000 explicitly
-        with option_context("compute.shortcut_limit", 1000):
-            pdf = pd.DataFrame(
-                {
-                    "A": [1, -2, 3, -4, 5] * 300,
-                    "B": [1.0, -2, 3, -4, 5] * 300,
-                    "C": [-6.0, -7, -8, -9, 10] * 300,
-                    "D": [True, False, True, False, False] * 300,
-                },
-                index=range(10, 15001, 10),
-            )
-            # TODO(SPARK-45228): Update `test_axis_on_dataframe` when Pandas regression is fixed
-            # There is a regression in Pandas 2.1.0,
-            # so we should manually cast to float until the regression is fixed.
-            # See https://github.com/pandas-dev/pandas/issues/55194.
-            pdf = pdf.astype(float)
-            psdf = ps.from_pandas(pdf)
-            self.assert_eq(psdf.count(axis=1), pdf.count(axis=1))
-            self.assert_eq(psdf.var(axis=1), pdf.var(axis=1))
-            self.assert_eq(psdf.var(axis=1, ddof=0), pdf.var(axis=1, ddof=0))
-            self.assert_eq(psdf.std(axis=1), pdf.std(axis=1))
-            self.assert_eq(psdf.std(axis=1, ddof=0), pdf.std(axis=1, ddof=0))
-            self.assert_eq(psdf.max(axis=1), pdf.max(axis=1))
-            self.assert_eq(psdf.min(axis=1), pdf.min(axis=1))
-            self.assert_eq(psdf.sum(axis=1), pdf.sum(axis=1))
-            self.assert_eq(psdf.product(axis=1), pdf.product(axis=1))
-            self.assert_eq(psdf.kurtosis(axis=0), pdf.kurtosis(axis=0), almost=True)
-            self.assert_eq(psdf.kurtosis(axis=1), pdf.kurtosis(axis=1))
-            self.assert_eq(psdf.skew(axis=0), pdf.skew(axis=0), almost=True)
-            self.assert_eq(psdf.skew(axis=1), pdf.skew(axis=1))
-            self.assert_eq(psdf.mean(axis=1), pdf.mean(axis=1))
-            self.assert_eq(psdf.sem(axis=1), pdf.sem(axis=1))
-            self.assert_eq(psdf.sem(axis=1, ddof=0), pdf.sem(axis=1, ddof=0))
-
-            self.assert_eq(
-                psdf.count(axis=1, numeric_only=True), pdf.count(axis=1, numeric_only=True)
-            )
-            self.assert_eq(psdf.var(axis=1, numeric_only=True), pdf.var(axis=1, numeric_only=True))
-            self.assert_eq(
-                psdf.var(axis=1, ddof=0, numeric_only=True),
-                pdf.var(axis=1, ddof=0, numeric_only=True),
-            )
-            self.assert_eq(psdf.std(axis=1, numeric_only=True), pdf.std(axis=1, numeric_only=True))
-            self.assert_eq(
-                psdf.std(axis=1, ddof=0, numeric_only=True),
-                pdf.std(axis=1, ddof=0, numeric_only=True),
-            )
-            self.assert_eq(
-                psdf.max(axis=1, numeric_only=True),
-                pdf.max(axis=1, numeric_only=True).astype(float),
-            )
-            self.assert_eq(
-                psdf.min(axis=1, numeric_only=True),
-                pdf.min(axis=1, numeric_only=True).astype(float),
-            )
-            self.assert_eq(
-                psdf.sum(axis=1, numeric_only=True),
-                pdf.sum(axis=1, numeric_only=True).astype(float),
-            )
-            self.assert_eq(
-                psdf.product(axis=1, numeric_only=True),
-                pdf.product(axis=1, numeric_only=True).astype(float),
-            )
-            self.assert_eq(
-                psdf.kurtosis(axis=0, numeric_only=True),
-                pdf.kurtosis(axis=0, numeric_only=True),
-                almost=True,
-            )
-            self.assert_eq(
-                psdf.kurtosis(axis=1, numeric_only=True), pdf.kurtosis(axis=1, numeric_only=True)
-            )
-            self.assert_eq(
-                psdf.skew(axis=0, numeric_only=True),
-                pdf.skew(axis=0, numeric_only=True),
-                almost=True,
-            )
-            self.assert_eq(
-                psdf.skew(axis=1, numeric_only=True), pdf.skew(axis=1, numeric_only=True)
-            )
-            self.assert_eq(
-                psdf.mean(axis=1, numeric_only=True), pdf.mean(axis=1, numeric_only=True)
-            )
-            self.assert_eq(psdf.sem(axis=1, numeric_only=True), pdf.sem(axis=1, numeric_only=True))
-            self.assert_eq(
-                psdf.sem(axis=1, ddof=0, numeric_only=True),
-                pdf.sem(axis=1, ddof=0, numeric_only=True),
-            )
-
     def test_skew_kurt_numerical_stability(self):
         pdf = pd.DataFrame(
             {
@@ -267,186 +171,6 @@ class StatsTestsMixin:
         psdf = ps.from_pandas(pdf)
         self.assert_eq(psdf.skew(), pdf.skew(), almost=True)
         self.assert_eq(psdf.kurt(), pdf.kurt(), almost=True)
-
-    def test_dataframe_corr(self):
-        pdf = pd.DataFrame(
-            index=[
-                "".join(
-                    np.random.choice(
-                        list("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"), 10
-                    )
-                )
-                for _ in range(30)
-            ],
-            columns=list("ABCD"),
-            dtype="float64",
-        )
-        psdf = ps.from_pandas(pdf)
-
-        with self.assertRaisesRegex(ValueError, "Invalid method"):
-            psdf.corr("std")
-        with self.assertRaisesRegex(TypeError, "Invalid min_periods type"):
-            psdf.corr(min_periods="3")
-
-        for method in ["pearson", "spearman", "kendall"]:
-            self.assert_eq(psdf.corr(method=method), pdf.corr(method=method), check_exact=False)
-            self.assert_eq(
-                psdf.corr(method=method, min_periods=1),
-                pdf.corr(method=method, min_periods=1),
-                check_exact=False,
-            )
-            self.assert_eq(
-                psdf.corr(method=method, min_periods=3),
-                pdf.corr(method=method, min_periods=3),
-                check_exact=False,
-            )
-            self.assert_eq(
-                (psdf + 1).corr(method=method, min_periods=2),
-                (pdf + 1).corr(method=method, min_periods=2),
-                check_exact=False,
-            )
-
-        # multi-index columns
-        columns = pd.MultiIndex.from_tuples([("X", "A"), ("X", "B"), ("Y", "C"), ("Z", "D")])
-        pdf.columns = columns
-        psdf.columns = columns
-
-        for method in ["pearson", "spearman", "kendall"]:
-            self.assert_eq(psdf.corr(method=method), pdf.corr(method=method), check_exact=False)
-            self.assert_eq(
-                psdf.corr(method=method, min_periods=1),
-                pdf.corr(method=method, min_periods=1),
-                check_exact=False,
-            )
-            self.assert_eq(
-                psdf.corr(method=method, min_periods=3),
-                pdf.corr(method=method, min_periods=3),
-                check_exact=False,
-            )
-            self.assert_eq(
-                (psdf + 1).corr(method=method, min_periods=2),
-                (pdf + 1).corr(method=method, min_periods=2),
-                check_exact=False,
-            )
-
-        # test with identical values
-        pdf = pd.DataFrame(
-            {
-                "a": [0, 1, 1, 1, 0],
-                "b": [2, 2, -1, 1, np.nan],
-                "c": [3, 3, 3, 3, 3],
-                "d": [np.nan, np.nan, np.nan, np.nan, np.nan],
-            }
-        )
-        psdf = ps.from_pandas(pdf)
-
-        for method in ["pearson", "spearman", "kendall"]:
-            self.assert_eq(psdf.corr(method=method), pdf.corr(method=method), check_exact=False)
-            self.assert_eq(
-                psdf.corr(method=method, min_periods=1),
-                pdf.corr(method=method, min_periods=1),
-                check_exact=False,
-            )
-            self.assert_eq(
-                psdf.corr(method=method, min_periods=3),
-                pdf.corr(method=method, min_periods=3),
-                check_exact=False,
-            )
-
-    def test_series_corr(self):
-        pdf = pd.DataFrame(
-            index=[
-                "".join(
-                    np.random.choice(
-                        list("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"), 10
-                    )
-                )
-                for _ in range(30)
-            ],
-            columns=list("ABCD"),
-            dtype="float64",
-        )
-        pser1 = pdf.A
-        pser2 = pdf.B
-        psdf = ps.from_pandas(pdf)
-        psser1 = psdf.A
-        psser2 = psdf.B
-
-        with self.assertRaisesRegex(ValueError, "Invalid method"):
-            psser1.corr(psser2, method="std")
-        with self.assertRaisesRegex(TypeError, "Invalid min_periods type"):
-            psser1.corr(psser2, min_periods="3")
-
-        for method in ["pearson", "spearman", "kendall"]:
-            self.assert_eq(
-                psser1.corr(psser2, method=method),
-                pser1.corr(pser2, method=method),
-                almost=True,
-            )
-            self.assert_eq(
-                psser1.corr(psser2, method=method, min_periods=1),
-                pser1.corr(pser2, method=method, min_periods=1),
-                almost=True,
-            )
-            self.assert_eq(
-                psser1.corr(psser2, method=method, min_periods=3),
-                pser1.corr(pser2, method=method, min_periods=3),
-                almost=True,
-            )
-            self.assert_eq(
-                (psser1 + 1).corr(psser2 - 2, method=method, min_periods=2),
-                (pser1 + 1).corr(pser2 - 2, method=method, min_periods=2),
-                almost=True,
-            )
-
-        # different anchors
-        psser1 = ps.from_pandas(pser1)
-        psser2 = ps.from_pandas(pser2)
-
-        with self.assertRaisesRegex(ValueError, "Cannot combine the series or dataframe"):
-            psser1.corr(psser2)
-
-        for method in ["pearson", "spearman", "kendall"]:
-            with ps.option_context("compute.ops_on_diff_frames", True):
-                self.assert_eq(
-                    psser1.corr(psser2, method=method),
-                    pser1.corr(pser2, method=method),
-                    almost=True,
-                )
-                self.assert_eq(
-                    psser1.corr(psser2, method=method, min_periods=1),
-                    pser1.corr(pser2, method=method, min_periods=1),
-                    almost=True,
-                )
-                self.assert_eq(
-                    psser1.corr(psser2, method=method, min_periods=3),
-                    pser1.corr(pser2, method=method, min_periods=3),
-                    almost=True,
-                )
-                self.assert_eq(
-                    (psser1 + 1).corr(psser2 - 2, method=method, min_periods=2),
-                    (pser1 + 1).corr(pser2 - 2, method=method, min_periods=2),
-                    almost=True,
-                )
-
-    def test_cov_corr_meta(self):
-        # Disable arrow execution since corr() is using UDT internally which is not supported.
-        with self.sql_conf({SPARK_CONF_ARROW_ENABLED: False}):
-            pdf = pd.DataFrame(
-                {
-                    "a": np.array([1, 2, 3], dtype="i1"),
-                    "b": np.array([1, 2, 3], dtype="i2"),
-                    "c": np.array([1, 2, 3], dtype="i4"),
-                    "d": np.array([1, 2, 3]),
-                    "e": np.array([1.0, 2.0, 3.0], dtype="f4"),
-                    "f": np.array([1.0, 2.0, 3.0]),
-                    "g": np.array([True, False, True]),
-                    "h": np.array(list("abc")),
-                },
-                index=pd.Index([1, 2, 3], name="myindex"),
-            )
-            psdf = ps.from_pandas(pdf)
-            self.assert_eq(psdf.corr(), pdf.corr(numeric_only=True), check_exact=False)
 
     def test_stats_on_boolean_dataframe(self):
         pdf = pd.DataFrame({"A": [True, False, True], "B": [False, False, True]})
@@ -588,7 +312,7 @@ class StatsTests(StatsTestsMixin, PandasOnSparkTestCase, SQLTestUtils):
 
 if __name__ == "__main__":
     import unittest
-    from pyspark.pandas.tests.test_stats import *  # noqa: F401
+    from pyspark.pandas.tests.computation.test_stats import *  # noqa: F401
 
     try:
         import xmlrunner
